backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
google-generativeai
uuid
pillow
replicate
fal-client
requests
